+++ app/auth/LoginResponse.scala
@@ -1,9 +1,10 @@
 package auth
 
+import auth.entity.Token
 import io.circe.Json
 import io.circe.syntax._
 
-case class TokenResponse(token: Token) {
+case class LoginResponse(token: Token) {
 
   def json: Json = Json.obj(
     "token" -> token.value.asJson
