+++ app/auth/entity/Token.scala
@@ -1,3 +1,3 @@
-package auth
+package auth.entity
 
 case class Token(value: String)
