+++ conf/routes
@@ -7,8 +7,8 @@
 GET     /                           controllers.HomeController.index
 GET     /health                     controllers.HealthController.check
 
-POST    /auth/login                 auth.AuthController.login
-POST    /auth/create                auth.AuthController.create
+POST    /auth/login                 auth.LoginController.login
+POST    /auth/create                auth.CreateUserController.create
 
 GET     /cafes                      cafe.CafeController.all
 GET     /cafe/:id                   cafe.CafeController.findById(id: Long)
